tasks.py: merge update-hound.py script
